Setup ajalt.clikt
