--- server/app/src/main/resources/application.conf
@@ -1,18 +1,0 @@
-ktor {
-    development = true
-    deployment {
-        port = "5000",
-        host = "0.0.0.0"
-    }
-    application {
-        modules = [ pos.ambrosia.AppKt.module ]
-    }
-    database {
-        path = ${HOME}/.Ambrosia-POS/ambrosia.db
-    }
-}
-jwt {
-    issuer = "http://0.0.0.0:3000/"
-    audience = "http://0.0.0.0:5000/"
-    realm = "Acces to 'Ambrosia-POS'"
-}
